use "unsigned int" when comparing to "unsigned int"
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@4331 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Pixes/pix_frei0r.cpp
@@ -418,7 +418,7 @@ void pix_frei0r :: parmMess(int key, int argc, t_atom *argv){
     error("negative parameterIDs not allowed");
     return;
   }
-  if(key>=m_plugin->m_parameterNames.size()) {
+  if(static_cast<unsigned int>(key)>=m_plugin->m_parameterNames.size()) {
     error("parameterID out of bounds");
     return;
   }
